% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advsearch.R
\name{get_advsearch}
\alias{get_advsearch}
\title{Retrieve BGG Advanced Search Results}
\usage{
get_advsearch(params)
}
\arguments{
\item{params}{A named list containing the parameters for the search.}
}
\value{
A data.frame containing the query results.
}
\description{
Retrieve BGG Advanced Search Results
}
\examples{
get_advsearch(list("q" = "Catan"))
}
