% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get Comments From BGG Game Page}
\usage{
get_comments(id)
}
\arguments{
\item{id}{character vector of BGG game IDs}
}
\value{
data frame of comments
}
\description{
Get Comments From BGG Game Page
}
\examples{

}
