% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments_page}
\alias{get_comments_page}
\title{Get Comments From BGG Game Page}
\usage{
get_comments_page(id, page)
}
\arguments{
\item{id}{character vector of BGG game IDs}

\item{page}{page number}
}
\value{
data frame of comments
}
\description{
Get Comments From BGG Game Page
}
\examples{

}
