% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments_search_page}
\alias{get_comments_search_page}
\title{Get comments for a given page of search results}
\usage{
get_comments_search_page(params, page)
}
\arguments{
\item{params}{parameters for advanced search}

\item{page}{page of advanced search}
}
\value{
comments dataframe for the search results
}
\description{
Get comments for a given page of search results
}
