% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forum.R
\name{get_forum}
\alias{get_forum}
\title{Retrieve BGG Advanced Search Results}
\usage{
get_forum(forum_id)
}
\arguments{
\item{forum_id}{The forum to retrieve.}
}
\value{
A data.frame containing the query results.
}
\description{
Retrieve BGG Advanced Search Results
}
\examples{

}
