% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forum.R
\name{get_forum_page}
\alias{get_forum_page}
\title{Retrieve Page of Advanced Search Results}
\usage{
get_forum_page(forum_id, page)
}
\arguments{
\item{forum_id}{The forum to retrieve.}

\item{page}{The page number  to retrieve.}
}
\value{
A data.frame containing the query results.
}
\description{
Retrieve Page of Advanced Search Results
}
\examples{
get_forum_page(forum_id = "8", page = "1")
}
