% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_games.R
\name{parse_games}
\alias{parse_games}
\title{Get Data From BGG Game Page}
\usage{
parse_games(id)
}
\arguments{
\item{id}{character vector of BGG game IDs}
}
\value{
data frame of basic information with one line for each game
}
\description{
Get Data From BGG Game Page
}
\examples{
parse_games(c('1', '42'))
}
